requirements.txt
face_recognition
opencv-python
requests
python-dotenv
basic-ftp
